Separate & move DropDown component styles to the assets folder
