+++ src/assets/styles/atoms/Container.js
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
-import { media, Solid } from '../index';
+import { media } from '../../assets/styles/variables/media';
 
 const Container = styled.div`
   max-width: 380px;
-  padding: 60px 40px;
-  border: 1px solid ${Solid.BORDER_GREY};
-  box-shadow: 0px 0px 25px rgba(0, 0, 0, 0.1);
+  border: 1px solid ${props => props.theme.color.grey.border};
   border-radius: 6px;
+  padding: 60px 40px;
+  box-shadow: 0px 0px 25px rgba(0, 0, 0, 0.1);
 
   @media ${media.tablet} {
     width: 340px;
